Fix AJP warning and remove patch
The upstream ebuild for nginx-1.9.4 still has the fatal warning about
AJP support not compiling.  It does, it works, it's nice that upstream
has an nginx_modules_http_ajp USE flag, but the warning is not necessary
since it doesn't seem to be true at the moment.

The changes to the nginx-clojure's module's initial hardwired config to
make it more compatible with Gentoo are (seemingly) no longer necessary.
Remove the patch from the ebuild, but keep the file around until
nginx-1.9.3-r1.ebuild (really, the dependence on commit
17c98149cf2e91a36754d34208aba6d47af56afd of nginx-clojure) goes away.