Firefox 46.0_beta11 bump
This is a literal rename bump from the mozilla overlay's
www-client/firefox-46.0_beta6 ebuild.

Merges, but still needs the GTK+ 3.20 fix, which I haven't adapted to
any of the .*_beta\d+ tags yet.  Probably in mozilla-central now, but
haven't found the commit yet, and need to get it cleanly applied against
a release tag.