Add 'ssl' USE flag to pjproject
OK, if you've already seen the ebuild and are wondering why no
src_configure-time options are being specified when clearly the
aconfigure.ac file suggests that --enable-ssl is an option, I can
happily tell you that you're right: aconfigure.ac does support behavior
for explicitly specifying that SSL/TLS support be enabled.  The problem
is that the behavior seems to be that the aconfigure script *ignores*
the SSL option and explicitly *disables* OpenSSL support whenever it's
passed in.  There's some logic in the aconfigure.ac/aconfigure files
that even has some info echoed when SSL support is disabled or when the
OpenSSL libraries cannot be found; they're passed over as well.
Something is clearly broken.
	However, the aconfigure script happily checks for the presence
of OpenSSL by default, and enables support if it finds it.  Thus, if
OpenSSL is already installed, SSL/TLS support will be enabled/built.
This does mean that if you have OpenSSL installed and don't want SSL/TLS
support built, tough.  Until aconfigure gets fixed, I have no interest
in trying to make a configure-time switch work.
	I suppose an interim fix would be to pass in --enable-ssl if the
'ssl' USE flag is not set, heh.