M libbitcoin-network, A libbitcoin-node
libbitcoin-network won't compile against libbitcoin, despite the fact
that it's built w/ a 32-bit library set.  It tries, instead, to link
against the 64-bit version of the library and fails.  Probably fixable,
not interested in fixing it right now.

Also, add libbitcoin-node.  Not sure if it's necessary for
libbitcoin-server, adding anyway.