Initial (rename) add of Asterisk 14.6.1.
This is a literal rename bump of the previous ebuild (14.2.0), with the
gentoo-asterisk-patchset tarball bumped to the version currently in the
Gentoo mirror set (should probably bump the older ebuilds too, would be
fun times seeing if it breaks anything).  No certainty whether this will
even get past src_prepare().