Update Asterisk(-14.6.1) patchset.
The patchset that is currently distributed by Gentoo is for a
significantly older version of Asterisk.  The changed version has no new
additions, and hopefully no change to functionality, but has been
applied against the closest version tag I could find, then rebased atop
the 14.6.1 tag.  We'll see if it works.