Add libxkbcommon live ebuild
Add a live (-9999-r1) ebuild for x11-libs/libxkbcommon.  The x11
overlay's ebuild is horribly out of date and written for autoconf, which
libxkbcommon no longer uses (it's since switched to meson).  This is a
first go at writing an ebuild that also casts of the xorg-2 eclass,
since it's not compatible with EABI 7.